+++ .github/workflows/e2e-tests.yml
@@ -29,7 +29,6 @@ jobs:
         category:
           - controller-manager
           - router
-          - router-plugins
           - gateway-api
           - gateway-inference-extension
     steps:
