+++ test/e2e/router/e2e_test.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/volcano-sh/kthena/test/e2e/framework"
 	routercontext "github.com/volcano-sh/kthena/test/e2e/router/context"
+	plugincontext "github.com/volcano-sh/kthena/test/e2e/router/router-plugins/context"
 	"github.com/volcano-sh/kthena/test/e2e/utils"
 )
 
@@ -69,8 +70,25 @@ func TestMain(m *testing.M) {
 		os.Exit(1)
 	}
 
+	pluginCtx := &plugincontext.PluginTestContext{
+		KubeClient:   testCtx.KubeClient,
+		KthenaClient: testCtx.KthenaClient,
+		Namespace:    testNamespace,
+	}
+	if err := pluginCtx.SetupPluginComponents(); err != nil {
+		fmt.Printf("Failed to setup plugin components: %v\n", err)
+		_ = testCtx.CleanupCommonComponents()
+		_ = testCtx.DeleteTestNamespace()
+		_ = framework.UninstallKthena(config.Namespace)
+		os.Exit(1)
+	}
+
 	// Run tests
 	code := m.Run()
+
+	if err := pluginCtx.CleanupPluginComponents(); err != nil {
+		fmt.Printf("Failed to cleanup plugin components: %v\n", err)
+	}
 
 	// Cleanup common components
 	if err := testCtx.CleanupCommonComponents(); err != nil {
