+++ test/e2e/router/plugins_helpers.go
@@ -1,0 +1,225 @@
+/*
+Copyright The Volcano Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package router
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"path/filepath"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+	backendmetrics "github.com/volcano-sh/kthena/pkg/kthena-router/backend/metrics"
+	plugincontext "github.com/volcano-sh/kthena/test/e2e/router/router-plugins/context"
+	"github.com/volcano-sh/kthena/test/e2e/utils"
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+)
+
+const (
+	pluginMockReplicaCount         = 3
+	leastRequestMaxWaitingRequests = 1
+	leastRequestLoadWaitTimeout    = 60 * time.Second
+)
+
+func listReadyMockPods(t *testing.T, kube kubernetes.Interface, namespace string) []corev1.Pod {
+	t.Helper()
+	ready := utils.ListReadyPodsByLabel(t, kube, namespace, "app="+plugincontext.AppLabel)
+	require.NotEmpty(t, ready, "no ready mock pods")
+	return ready
+}
+
+// deploySlowLatencyMockStack deploys the slow mock backend for least-latency e2e.
+// Recreate so a prior run cannot leave multiple ready slow pods.
+func deploySlowLatencyMockStack(t *testing.T, kube kubernetes.Interface, namespace string) {
+	t.Helper()
+	ctx := context.Background()
+	name := plugincontext.SlowMockDeploymentName
+
+	utils.DeleteDeploymentAndWait(t, kube, namespace, name, 2*time.Minute)
+
+	deployment := utils.LoadYAMLFromFile[appsv1.Deployment](filepath.Join(plugincontext.TestDataDir, "LLM-Mock-plugins-slow.yaml"))
+	deployment.Namespace = namespace
+	_, err := kube.AppsV1().Deployments(namespace).Create(ctx, deployment, metav1.CreateOptions{})
+	require.NoError(t, err)
+	t.Cleanup(func() {
+		_ = kube.AppsV1().Deployments(namespace).Delete(context.Background(), name, metav1.DeleteOptions{})
+	})
+	utils.WaitForDeploymentReady(t, ctx, kube, namespace, name, 1, 5*time.Minute)
+	require.Eventually(t, func() bool {
+		return len(utils.ListReadyPodsByLabel(t, kube, namespace, "app="+plugincontext.SlowMockAppLabel)) == 1
+	}, 2*time.Minute, 2*time.Second, "expected exactly one ready slow mock pod")
+}
+
+func waitForSchedulerPluginInMetrics(t *testing.T, metricsURL, pluginName, pluginType string) {
+	t.Helper()
+	require.Eventually(t, func() bool {
+		metricsData, err := backendmetrics.ParseMetricsURL(metricsURL)
+		if err != nil {
+			return false
+		}
+		return utils.GetHistogramCount(metricsData, "kthena_router_scheduler_plugin_duration_seconds", map[string]string{
+			"plugin": pluginName,
+			"type":   pluginType,
+		}) > 0
+	}, 30*time.Second, time.Second)
+}
+
+type routerPodMetricsSnapshot struct {
+	RequestWaitingNum float64
+	RequestRunningNum float64
+}
+
+func fetchRouterPodMetricsViaDebug(t *testing.T, debugBaseURL string, pod corev1.Pod) (routerPodMetricsSnapshot, bool) {
+	t.Helper()
+	url := fmt.Sprintf("%s/debug/config_dump/namespaces/%s/pods/%s", debugBaseURL, pod.Namespace, pod.Name)
+	client := &http.Client{Timeout: 5 * time.Second}
+	resp, err := client.Get(url)
+	if err != nil {
+		return routerPodMetricsSnapshot{}, false
+	}
+	defer resp.Body.Close()
+	if resp.StatusCode != http.StatusOK {
+		return routerPodMetricsSnapshot{}, false
+	}
+	var parsed struct {
+		Metrics *struct {
+			RequestWaitingNum float64 `json:"requestWaitingNum"`
+			RequestRunningNum float64 `json:"requestRunningNum"`
+		} `json:"metrics"`
+	}
+	if err := json.NewDecoder(resp.Body).Decode(&parsed); err != nil || parsed.Metrics == nil {
+		return routerPodMetricsSnapshot{}, false
+	}
+	return routerPodMetricsSnapshot{
+		RequestWaitingNum: parsed.Metrics.RequestWaitingNum,
+		RequestRunningNum: parsed.Metrics.RequestRunningNum,
+	}, true
+}
+
+func waitForLeastRequestLoadSeparation(t *testing.T, kube kubernetes.Interface, kthenaNamespace string, busyPods []corev1.Pod, idlePod corev1.Pod, maxWaitingRequests int) {
+	t.Helper()
+	require.NotEmpty(t, busyPods)
+	require.Greater(t, maxWaitingRequests, 0)
+	threshold := float64(maxWaitingRequests)
+
+	routerPod := utils.GetRouterPod(t, kube, kthenaNamespace)
+	localPort := utils.AllocateLocalPort(t)
+	pf, err := utils.SetupPortForwardToPod(routerPod.Namespace, routerPod.Name, localPort, utils.RouterDebugPort)
+	require.NoError(t, err, "port-forward to router debug API")
+	defer pf.Close()
+
+	debugBaseURL := fmt.Sprintf("http://127.0.0.1:%s", localPort)
+	require.Eventually(t, func() bool {
+		allBusySaturated := true
+		for _, busyPod := range busyPods {
+			busy, okBusy := fetchRouterPodMetricsViaDebug(t, debugBaseURL, busyPod)
+			if !okBusy || busy.RequestWaitingNum < threshold {
+				allBusySaturated = false
+				break
+			}
+		}
+		idle, okIdle := fetchRouterPodMetricsViaDebug(t, debugBaseURL, idlePod)
+		if !okIdle {
+			return false
+		}
+		idleFree := idle.RequestWaitingNum < threshold
+		if allBusySaturated && idleFree {
+			for _, busyPod := range busyPods {
+				busy, _ := fetchRouterPodMetricsViaDebug(t, debugBaseURL, busyPod)
+				t.Logf("least-request load ready: busy %s waiting=%.0f running=%.0f",
+					busyPod.Name, busy.RequestWaitingNum, busy.RequestRunningNum)
+			}
+			t.Logf("least-request load ready: idle %s waiting=%.0f running=%.0f",
+				idlePod.Name, idle.RequestWaitingNum, idle.RequestRunningNum)
+		}
+		return allBusySaturated && idleFree
+	}, leastRequestLoadWaitTimeout, 2*time.Second,
+		"all busy pods should have request_waiting >= %d and idle pod %s should have request_waiting < %d",
+		maxWaitingRequests, idlePod.Name, maxWaitingRequests)
+}
+
+const (
+	schedulerOnlyPrefixCache = `scheduler:
+  pluginConfig:
+  - name: prefix-cache
+    args:
+      blockSizeToHash: 64
+      maxBlocksToMatch: 128
+      maxHashCacheSize: 50000
+      topKMatches: 5
+  plugins:
+    Filter:
+      enabled: []
+    Score:
+      enabled:
+        - name: prefix-cache
+          weight: 1`
+
+	schedulerOnlyLeastRequest = `scheduler:
+  pluginConfig:
+  - name: least-request
+    args:
+      maxWaitingRequests: 1
+  plugins:
+    Filter:
+      enabled:
+        - least-request
+    Score:
+      enabled:
+        - name: least-request
+          weight: 1`
+
+	schedulerOnlyLeastLatency = `scheduler:
+  pluginConfig:
+  - name: least-latency
+    args:
+      TTFTTPOTWeightFactor: 0.5
+  plugins:
+    Filter:
+      enabled: []
+    Score:
+      enabled:
+        - name: least-latency
+          weight: 1`
+
+	schedulerOnlyLoraAffinity = `scheduler:
+  pluginConfig: []
+  plugins:
+    Filter:
+      enabled:
+        - lora-affinity
+    Score:
+      enabled:
+        - name: random
+          weight: 1`
+
+	schedulerOnlyRandom = `scheduler:
+  pluginConfig: []
+  plugins:
+    Filter:
+      enabled: []
+    Score:
+      enabled:
+        - name: random
+          weight: 1`
+)
