+++ test/e2e/router/router-plugins/testdata/ModelServer-plugins.yaml
